refactor: Extract error whiskers rendering to a shared folder
